New switch function

+++ net_dfir.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 clear
-export GREP_COLORS="01;35"
 
 RED='\033[31m'
 GREEN='\033[32m'
@@ -15,80 +14,57 @@ YELLOW_BG="\e[1;43m"
 RED_BG='\e[1;41m'
 DIV="=============================="
 
-
-BANNER="
-
-███╗   ██╗███████╗████████╗    ██████╗ ███████╗██╗██████╗ 
-████╗  ██║██╔════╝╚══██╔══╝    ██╔══██╗██╔════╝██║██╔══██╗
-██╔██╗ ██║█████╗     ██║       ██║  ██║█████╗  ██║██████╔╝
-██║╚██╗██║██╔══╝     ██║       ██║  ██║██╔══╝  ██║██╔══██╗
-██║ ╚████║███████╗   ██║       ██████╔╝██║     ██║██║  ██║
-╚═╝  ╚═══╝╚══════╝   ╚═╝       ╚═════╝ ╚═╝     ╚═╝╚═╝  ╚═╝
-"
-
 GEO_HIGHLIGHT="russia|iran|lithuania|china|cyprus|hong\skong|united\sarab\semirates|$"
 AGENT_HIGHLIGHT="python|curl|wget|$"
 HOST_HIGHLIGHT="python|simplehttp|$"
 FILE_HIGHLIGHT="\.[a-zA-Z]+|$"
 PORT_HIGHLIGHT="139|445|135|4444|9001|8080|8001|8000|3389$"
-OBJECT_HIGHLIGHT="\.zip|\.rar|\.exe|$"
+OBJECT_HIGHLIGHT="\.zip|\.rar|\.exe|\.dat|\.dll|$"
 
-OPTIONS=("PCAP")
-SWITCHES=("-r", "--ipfix", "--cidr", "-d")
-OPTION_IDX=0
-NEXT_ARG_VARIABLE=""
-REMOVAL_REQUIRED=0
+PCAP=""
+EXPORT=false
 
-for arg in "$@"; do
-	if [ ! -z "$NEXT_ARG_VARIABLE" ]; then
-		eval "${NEXT_ARG_VARIABLE}=\"$arg\""
-		NEXT_ARG_VARIABLE=""
-		continue
-	fi
-
-	if [[ ${SWITCHES[@]} =~ $arg ]]; then
-		case $arg in
-		"-r")
-			REMOVAL_REQUIRED=1
-			;;
-		"--ipfix")
-			NEXT_ARG_VARIABLE="NETFLOW_FILE"
-			;;
-		"--cidr")
-			NEXT_ARG_VARIABLE="CIDR"
-			;;
-		"-d")
-			NEXT_ARG_VARIABLE="DIR"
-			;;
-				
-		esac
-		continue
-	else
-
-		case $OPTION_IDX in
-			0)
-				PCAP="$(realpath $arg)"
-				TIMESTAMP="$(date +%y%m%d%H%M%S)"
-				if [ -z "$DIR" ]; then
-					DIR="$PCAP-$TIMESTAMP"
-				fi
-				;;
-		esac
-	
-		((OPTION_IDX++))
-	fi
-
-DIR="$(realpath $DIR)"
-
+while getopts "r:e" opt; do
+    case $opt in
+        "r")
+            PCAP="$OPTARG"
+            ;;
+        "e")
+            EXPORT=true
+            ;;
+    esac
 done
 
-if [ -z "$NETFLOW_FILE" ]; then
-	NETFLOW_FILE="$DIR/netflow.silk"
-fi
+banner() {
+    echo
+    echo -e "\e[31m███╗   ██╗███████╗████████╗    ██████╗ ███████╗██╗██████╗ \e[0m"
+    echo -e "\e[33m████╗  ██║██╔════╝╚══██╔══╝    ██╔══██╗██╔════╝██║██╔══██╗\e[0m"
+    echo -e "\e[32m██╔██╗ ██║█████╗     ██║       ██║  ██║█████╗  ██║██████╔╝\e[0m"
+    echo -e "\e[34m██║╚██╗██║██╔══╝     ██║       ██║  ██║██╔══╝  ██║██╔══██╗\e[0m"
+    echo -e "\e[35m██║ ╚████║███████╗   ██║       ██████╔╝██║     ██║██║  ██║\e[0m"
+    echo -e "\e[36m╚═╝  ╚═══╝╚══════╝   ╚═╝       ╚═════╝ ╚═╝     ╚═╝╚═╝  ╚═╝\e[0m"
+    echo
+    echo -e "${RED}Perform baselining and analysis on network captures.${RESET}"
+    echo
+}
 
-# create the log dir if it doesn't exist
-if [ ! -d "$DIR" ]; then
-	mkdir $DIR
+help() {
+    banner
+    echo "Usage: $0 [-r <pcap_file>] [-e]"
+    echo
+    echo "Options:"
+    echo "  -r <pcap_file>      Specify the input PCAP file for analysis (required)"
+    echo "  -e                  Export files detected in packet streams (optional)"
+    echo
+    echo "Example:"
+    echo "  $0 -r capture.pcap -e"
+    exit
+}
+
+if [[ $PCAP == *.pcap || $PCAP == *.cap ]]; then
+    DIR="$(realpath $PCAP.zeek)"
+else
+    help
 fi
 
 log_header() {
@@ -109,13 +85,7 @@ log_value() {
 	echo -e "${RED}${BOLD}[+] ${BLUE}${title}: ${RESET}${val}"
 }
 
-log_banner() {
-	echo "$BANNER"
-}
-
 zeek_create() {
-    echo $(log_header "Zeek")
-	echo
     echo "Generating Zeek Output..."
     echo
 
@@ -127,79 +97,23 @@ zeek_create() {
     mv *.log $DIR
 }
 
-netflow_create() {
-	echo $(log_header "NETFLOW")
-	echo
-
-	rwp2yaf2silk --in $PCAP --out $NETFLOW_FILE
-}
-
-netflow_all_tcp_ports(){
-	echo $(log_header "NETFLOW ALL TCP DESTINATION PORTS")
-	echo
-
-	# if the cidr is set, use it
-	CIDR_FILTER=""
-	if [ ! -z "$CIDR" ]; then
-		CIDR_FILTER="--dcidr=$CIDR"
-	fi
-
-	rwfilter --type=all --proto=6 $CIDR_FILTER --flags-initial=S/SA --pass=stdout $NETFLOW_FILE | rwsort --field=stime | rwstats --fields=dport --count=1000
-}
-
-netflow_all_udp_ports(){
-	echo $(log_header "NETFLOW ALL UDP DESTINATION PORTS")
-	echo
-
-	# if the cidr is not set exit with an error
-	CIDR_FILTER=""
-	if [ -z "$CIDR" ]; then
-		echo $(log_value "CIDR" "CIDR is required for this analysis")
-		return
-	fi
-
-	CIDR_FILTER="--dcidr=$CIDR"
-
-	rwfilter --type=all --proto=17 $CIDR_FILTER --pass=stdout $NETFLOW_FILE | rwsort --field=stime | rwstats --fields=dport --values=records --threshold=3
-}
-
-netflow_excessive_unique_requests(){
-	echo $(log_header "NETFLOW EXCESSIVE UNIQUE REQUESTS (<0.5s)")
-	echo
-
-	# if the cidr is set, use it
-	CIDR_FILTER=""
-	if [ ! -z "$CIDR" ]; then
-		CIDR_FILTER="--dcidr=$CIDR"
-	fi
-
-	# Shows all source IPs that have made more than 10 unique requests to the same destination IP and port in less than 0.5 seconds
-	rwfilter --type=all --proto=0-255 --pass=stdout $NETFLOW_FILE | rwfilter --duration=0.0-0.5 - --pass=stdout | rwsort --field=stime | rwuniq --fields=sip,dip,dport --values=distinct:sport --threshold=distinct:sport=10
-}
-
-netflow_all_ip_addresses(){
-	echo $(log_header "NETFLOW ALL IP ADDRESSES")
-	echo
-
-	# Shows the count of all the records to each IP address
-	(rwfilter --type=all --proto=0-255 --pass=stdout $NETFLOW_FILE | rwsort --field=stime | rwcut --fields=sip &&
-        rwfilter --type=all --proto=0-255 --pass=stdout $NETFLOW_FILE | rwsort --field=stime | rwcut --fields=dip ) | sort | uniq -c | egrep -v "1\s+[sd]IP" | sort -nr
-}
-
 mmdb_check() {
     if [ ! -x "$(command -v mmdblookup)" ]; then
-        echo -e "${RED}Installing mmdblookup...${RESET}"
-        sudo apt install libmaxminddb0 libmaxminddb-dev mmdb-bin geoipupdate -y
+        echo "Installing mmdblookup..."
+        sudo apt install libmaxminddb0 libmaxminddb-dev mmdb-bin geoipupdate -y &>/dev/null
     fi
 
     if [ ! -f "/tmp/geo-city.mmdb" ]; then
-        echo -e "${RED}Installing mmdb database...${RESET}"
+        echo "Installing mmdb database..."
         wget https://git.io/GeoLite2-City.mmdb -O /tmp/geo-city.mmdb &>/dev/null
     fi
 }
 
 download_threat_intel() {
-    wget https://raw.githubusercontent.com/stamparm/ipsum/refs/heads/master/ipsum.txt -O /tmp/ipsum.txt &>/dev/null
+    if [ ! -f "/tmp/ipsum.txt" ]; then
+        echo "Downloading threat intel database..."
+        wget https://raw.githubusercontent.com/stamparm/ipsum/refs/heads/master/ipsum.txt -O /tmp/ipsum.txt &>/dev/null
+    fi
 }
 
 get_environment() {
@@ -382,6 +296,7 @@ get_anomolous_dc_activity() {
         list+="$occurences,$src_ip_stat -> $dst_ip_stat,:$port,$country\n"
     done <<< "$traffic"
 
+    echo $(log_value "Outgoing Traffic")
     echo -e "$list" | column -t -s "," | sort -k 1 -n -r
 
     echo
@@ -543,23 +458,10 @@ get_smb_objects() {
     echo
 }
 
-zeek_remove() {
-  if [ $REMOVAL_REQUIRED -eq 1 ]; then
-    rm -rf $DIR
-  fi
-}
-
 execute_all() {
-	  log_banner
+	banner
     mmdb_check
     download_threat_intel
-    
-    # netflow_create
-    # netflow_all_tcp_ports
-    # netflow_all_udp_ports
-    # netflow_all_ip_addresses
-    # netflow_excessive_unique_requests
-    
     zeek_create
     get_environment
     get_active_directory
@@ -571,10 +473,11 @@ execute_all() {
     get_user_agents
     get_server_hosts
     get_uris
-    get_http_objects
-    get_smb_objects
-    get_external_connections
-    zeek_remove
+
+    if [ $EXPORT == true ]; then
+        get_http_objects
+        get_smb_objects
+    fi
 }
 
 execute_all
